Renamed guTracking to acquisitions across the codebase.
